- replaced etl::string_view with etl::string<SSIZE> - replaced .data() method with .c_str() in functions which expect c type strings -  added debugging flags for both release and debug

+++ Core/Src/main.c
@@ -319,7 +319,7 @@ static void MX_GPIO_Init(void)
 int _write(int file, char* ptr, int len)
 {
     (void) file;
-    HAL_UART_Transmit(&huart2, (const uint8_t*) ptr, len, 100);
+    HAL_UART_Transmit(&huart2, (const uint8_t*) ptr, (uint16_t) len, 100);
     return len;
 }
 /* USER CODE END 4 */
